% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{bind_between}
\alias{bind_between}
\title{function forcing value to fall between min and max}
\usage{
bind_between(x, min, max)
}
\arguments{
\item{x}{the values to be bound}

\item{min}{lower boundary}

\item{max}{upper boundary}
}
\description{
function forcing value to fall between min and max
}
\keyword{internal}

