% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{classes}
\alias{classes}
\title{function to get classes from e.g. lists}
\usage{
classes(x)
}
\arguments{
\item{x}{list to get classes for}
}
\description{
function to get classes from e.g. lists
}
\keyword{internal}

