% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffrproject.R
\docType{class}
\name{diffrproject}
\alias{diffrproject}
\title{class for diffrproject}
\format{\code{\link{R6Class}} object.}
\usage{
diffrproject
}
\value{
Object of \code{\link{R6Class}}
}
\description{
class for diffrproject
}
\keyword{data}

