% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dim1}
\alias{dim1}
\title{get first dimension or length of object}
\usage{
dim1(x)
}
\arguments{
\item{x}{object, matrix, vector, data.frame, ...}
}
\description{
get first dimension or length of object
}
\keyword{internal}

