% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\docType{data}
\name{dp_storage}
\alias{dp_storage}
\title{storage for internals}
\format{An object of class \code{environment} of length 0.}
\usage{
dp_storage
}
\description{
storage for internals
}
\keyword{datasets}

