% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_text.R
\docType{class}
\name{dp_text}
\alias{dp_text}
\title{text class}
\format{\code{\link{R6Class}} object.}
\usage{
dp_text
}
\value{
Object of \code{\link{R6Class}} with method(s) for bot permission checking.
}
\description{
text class
}
\section{Fields}{

\describe{
\item{\code{text}}{a single character string / character  vector of length one}

\item{\code{tokens}}{a data frame with a character vector representing the the text as tokens,
and two integer vectors capturing the span of the token measured in
characters starting from beginning of text}

\item{\code{file}}{path to a file from which text was read in}

\item{\code{encoding}}{encoding to be assumed for the text}

\item{\code{soourcetype}}{list of logicals that captures where the text came from when
initializing object: text or file}
}}
\section{Methods}{

\describe{
 \item{
   \code{
   initialize(domain, text) }}{
   Method called when initialising object via
   \code{robotstxt$new()}. Needs either \code{text} or \code{domain} to be
   present at initialization. If only domain is supplied -- should be seen as
   default -- than the robots.txt file for that domain will be downloaded. If
   \code{text} is supplied as well, nothing will be downloaded. If only
   \code{text} is supplied than domain is set to '???'.
 }
 \item{\code{
   check( path="/", bot="*" ) }}{
   Method for checking whether or not paths are allowed to be accessed by a bot
 }
}
}
\examples{
mytext <- dp_text$new("Hallo World")
mytext$show_text()

mytext <- dp_text$new(c("Hallo","World"))
mytext$show_text()

}
\keyword{data}

