% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tools.R
\name{dp_text_base_data}
\alias{dp_text_base_data}
\title{function providing basic information on texts within diffrproject}
\usage{
dp_text_base_data(dp)
}
\arguments{
\item{dp}{a diffrproject object}
}
\description{
function providing basic information on texts within diffrproject
}

