% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span_tools.R
\name{expand_spans}
\alias{expand_spans}
\title{expanding time spans to sequernce of days}
\usage{
expand_spans(x)
}
\arguments{
\item{x}{a matrix like object with dates; column 1: start dates; column 2: end dates}
}
\description{
expanding time spans to sequernce of days
}

