% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_vector_element}
\alias{get_vector_element}
\title{function that extracts elements from vector}
\usage{
get_vector_element(vec, length = NULL, from = NULL, to = NULL)
}
\arguments{
\item{vec}{the chars field}

\item{length}{number of elements to be returned}

\item{from}{first element to be returned}

\item{to}{last element to be returned}
}
\description{
function that extracts elements from vector
}
\keyword{internal}

