% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_between}
\alias{is_between}
\title{function that checks is values are in between values}
\usage{
is_between(x, y, z)
}
\arguments{
\item{x}{input vector}

\item{y}{lower bound}

\item{z}{upper bound}
}
\description{
function that checks is values are in between values
}
\keyword{internal}

