% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{load_into}
\alias{load_into}
\title{function that loads saved rtext}
\usage{
load_into(save_file)
}
\arguments{
\item{save_file}{a saved rtext object in Rdata format}
}
\description{
function that loads saved rtext
}

