% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span_tools.R
\name{mfts_match_dates_to_id}
\alias{mfts_match_dates_to_id}
\title{matching ids to dates}
\usage{
mfts_match_dates_to_id(full_sequence, df)
}
\arguments{
\item{df}{a data frame containing span inforamtion with}

\item{full_day_sequence}{a sequence of days span ids should be matched to}
}
\description{
function matching a sequence of dates and a data frame of id, start, end
}

