% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\name{plot.rtext}
\alias{plot.rtext}
\title{function for plotting rtext}
\usage{
\method{plot}{rtext}(rtext, char_data = NULL, lines = NULL,
  col = "#ED4C4CA0", xlim = c(0, (ceiling(max(x)/10^nchar(max(x)) * 10)) *
  (10^nchar(max(x))/10)), ylim = c(0, max(y) + 1), ...)
}
\arguments{
\item{rtext}{object of class rtext}

\item{char_data}{name of a variable/column to be marked}

\item{lines}{vector of integer listing the lines to be plottted}

\item{col}{color of the char_data variable to be highlighted}

\item{xlim}{xlim passed through to initial plot}

\item{ylim}{ylim passed through to initial plot}

\item{...}{further parameters passed through to initial plot}
}
\description{
function for plotting rtext
}

