% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\docType{data}
\name{prometheus_early}
\alias{prometheus_early}
\title{prometheus early version}
\format{
An object of class \code{character} of length 1.
}
\source{
https://de.wikisource.org/w/index.php?title=Prometheus_(Gedicht,_fr%C3%BChe_Fassung)&oldid=2105646
}
\usage{
prometheus_early
}
\description{
prometheus early version
}
\keyword{datasets}
