% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\docType{data}
\name{prometheus_late}
\alias{prometheus_late}
\title{prometheus late version}
\format{
An object of class \code{character} of length 1.
}
\source{
https://de.wikisource.org/w/index.php?title=Prometheus_(Gedicht,_sp%C3%A4te_Fassung)&oldid=1479587
}
\usage{
prometheus_late
}
\description{
prometheus late version
}
\keyword{datasets}
