% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_rtext_extended.r
\docType{class}
\name{R6_rtext_extended}
\alias{R6_rtext_extended}
\title{extended R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
R6_rtext_extended
}
\value{
Object of \code{\link{R6Class}}
}
\description{
extended R6 class
}
\seealso{
\code{\link{rtext}}
}
\keyword{data}

