% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rbind_fill}
\alias{rbind_fill}
\title{function for binding data.frames even if names do not match}
\usage{
rbind_fill(df1 = data.frame(), df2 = data.frame())
}
\arguments{
\item{df1}{first data.frame to rbind}

\item{df2}{second data.frame to rbind}
}
\description{
function for binding data.frames even if names do not match
}
\keyword{internal}

