% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_base.R
\docType{class}
\name{rtext_base}
\alias{rtext_base}
\title{rtext_base : basic workhorse for rtext}
\format{\code{\link{R6Class}} object.}
\usage{
rtext_base
}
\value{
Object of \code{\link{R6Class}}
}
\description{
rtext_base : basic workhorse for rtext
}
\seealso{
\code{\link{rtext}}
}
\keyword{data}
