% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\name{rtext_get_character}
\alias{rtext_get_character}
\title{function to get text from rtext object}
\usage{
rtext_get_character(chars, length = 100, from = NULL, to = NULL)
}
\arguments{
\item{chars}{the chars field}

\item{length}{number of characters to be returned}

\item{from}{first character to be returned}

\item{to}{last character to be returned}
}
\description{
function to get text from rtext object
}
\keyword{internal}
