% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rtext_hash}
\alias{rtext_hash}
\title{function to get hash for R objects}
\usage{
rtext_hash(x)
}
\arguments{
\item{x}{the thing to hash}
}
\description{
function to get hash for R objects
}
\keyword{internal}

