% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tokenize.R
\docType{class}
\name{rtext_tokenize}
\alias{rtext_tokenize}
\title{R6 class - linking text and data}
\format{\code{\link{R6Class}} object.}
\usage{
rtext_tokenize
}
\value{
Object of \code{\link{R6Class}}
}
\description{
R6 class - linking text and data
}
\seealso{
\code{\link{rtext}}
}
\keyword{data}

