% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext.R
\docType{data}
\name{rtext_tokenizer}
\alias{rtext_tokenizer}
\title{list of ready to use functions for rtext initialization and tokenization}
\format{An object of class \code{list} of length 1.}
\usage{
rtext_tokenizer
}
\description{
list of ready to use functions for rtext initialization and tokenization
}
\keyword{datasets}

