% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\docType{data}
\name{rtext_tokenizer_list}
\alias{rtext_tokenizer_list}
\title{list of ready to use functions for rtext initialization and tokenization}
\format{An object of class \code{list} of length 3.}
\usage{
rtext_tokenizer_list
}
\description{
list of ready to use functions for rtext initialization and tokenization
}
\keyword{datasets}

