% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inheritance.R
\docType{data}
\name{sub_class}
\alias{sub_class}
\title{a sub_class that inherits}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sub_class
}
\description{
a sub_class that inherits
}
\keyword{datasets}

