% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inheritance.R
\docType{data}
\name{super_class}
\alias{super_class}
\title{a super class to inherit from}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
super_class
}
\description{
a super class to inherit from
}
\keyword{datasets}

