% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_file}
\alias{test_file}
\title{text function: wrapper for system.file() to access test files}
\usage{
test_file(x = NULL, pattern = NULL, full.names = FALSE)
}
\arguments{
\item{x}{name of the file}

\item{pattern}{pattern of file name}
}
\description{
text function: wrapper for system.file() to access test files
}
\keyword{internal}

