% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{text_collapse}
\alias{text_collapse}
\alias{text_collapse.data.frame}
\alias{text_collapse.default}
\alias{text_collapse.list}
\title{function for collapsing text vectors}
\usage{
text_collapse(x, ..., sep)

\method{text_collapse}{default}(x, ..., sep = "")

\method{text_collapse}{list}(x, ..., sep = c("", ""))

\method{text_collapse}{data.frame}(x, ..., sep = c("", "\\n"))
}
\arguments{
\item{x}{object to be collapsed}

\item{...}{additional passed on to specific methods (e.g. to paste within text_collapse.default())}

\item{sep}{separator between text parts}
}
\description{
function for collapsing text vectors

default method for text_collapse()

text_collapse() method for lists

text_collapse() method for data.frames
}

