% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{text_read}
\alias{text_read}
\title{helper function: to read in text}
\usage{
text_read(file, tokenize = NULL, encoding = "UTF-8")
}
\arguments{
\item{file}{name of the file to be read in}

\item{tokenize}{either
NULL so that no splitting is done;
a character to use to split text into parts;
or a function that does the splitting}

\item{encoding}{character encoding of file}
}
\description{
a wrapper to readLines() to make things more ordered and convenient
}

