% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{text_snippet}
\alias{text_snippet}
\title{helper function: retrieving text snippet}
\usage{
text_snippet(x, length = 500, from = NULL, to = NULL, coll = FALSE)
}
\arguments{
\item{x}{character vector to be snipped}

\item{length}{length of snippet}

\item{from}{starting character}

\item{to}{last character}

\item{coll}{should a possible vector x with length > 1 collapsed with newline
character as separator?}
}
\description{
function will give back snippets of length length for length-from or
length-to combinations but will give snippets of to-from length for parameter
from-to-combinations
}

