% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{text_tokenize.rtext}
\alias{text_tokenize.rtext}
\title{function tokenizing rtext objects}
\usage{
\method{text_tokenize}{rtext}(x, regex = NULL, ignore.case = FALSE,
  fixed = FALSE, perl = FALSE, useBytes = FALSE, non_token = FALSE)
}
\description{
function tokenizing rtext objects
}

