% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\name{vector_delete}
\alias{vector_delete}
\title{function used to delete parts from a vector}
\usage{
vector_delete(x, n = NULL, from = NULL, to = NULL)
}
\arguments{
\item{x}{input vector}

\item{n}{number of items to be deleted}

\item{from}{from which position onwards elements should be deleted}

\item{to}{up to which positions elements should be deleted}
}
\description{
function used to delete parts from a vector
}

