% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{which_token}
\alias{which_token}
\title{function returning index of spans that entail x}
\usage{
which_token(x, y1, y2)
}
\arguments{
\item{x}{position of the character}

\item{y1}{start position of the token}

\item{y2}{end position of the token}
}
\description{
function returning index of spans that entail x
}
\keyword{internal}
