% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_token_worker_user_interrupt}
\alias{which_token_worker_user_interrupt}
\title{function to check which chars belong to which token
takes a vector of xs to check if these lie between pairs of ys and if so
returning their index; assumes xs and ys are sorted; returns only the first
span index which enclosing the x}
\usage{
which_token_worker_user_interrupt(x, y1, y2)
}
\arguments{
\item{x}{whatever}

\item{y1}{whatever}

\item{y2}{whatever}
}
\description{
function to check which chars belong to which token
takes a vector of xs to check if these lie between pairs of ys and if so
returning their index; assumes xs and ys are sorted; returns only the first
span index which enclosing the x
}

