% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\name{plot.rtext}
\alias{plot.rtext}
\title{function for plotting rtext}
\usage{
\method{plot}{rtext}(x, y = NULL, lines = TRUE, what = NULL,
  col = "#ED4C4CA0", aggregate_function = function(x) {     any(!is.na(x))
  }, ...)
}
\arguments{
\item{x}{object of class rtext}

\item{y}{name of a variable/column to be marked}

\item{lines}{vector of integer listing the lines to be plottted}

\item{col}{color of the char_data variable to be highlighted}

\item{...}{further parameters passed through to initial plot}

\item{char_data}{name of a variable/column to be marked}
}
\description{
function for plotting rtext
}

