% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_tools.R
\name{plot.rtext}
\alias{plot.rtext}
\title{function for plotting rtext}
\usage{
\method{plot}{rtext}(x, y = NULL, lines = TRUE, col = "#ED4C4CA0",
  add = FALSE, ...)
}
\arguments{
\item{x}{object of class rtext}

\item{y}{char_data to be plotted}

\item{lines}{vector of integer listing the lines to be plottted}

\item{col}{color of the char_data variable to be highlighted}

\item{add}{add data to an already existing plot?}

\item{...}{further parameters passed through to initial plot}
}
\description{
function for plotting rtext
}

