% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_rtext_extended.r
\docType{class}
\name{R6_rtext_extended}
\alias{R6_rtext_extended}
\title{extended R6 class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}}
}
\description{
extended R6 class

extended R6 class
}
\seealso{
\code{\link{rtext}}
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get}{\code{R6_rtext_extended$get()}}
\item \href{#method-debug}{\code{R6_rtext_extended$debug()}}
\item \href{#method-ls}{\code{R6_rtext_extended$ls()}}
\item \href{#method-message}{\code{R6_rtext_extended$message()}}
\item \href{#method-warning}{\code{R6_rtext_extended$warning()}}
\item \href{#method-clone}{\code{R6_rtext_extended$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_rtext_extended$get(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug"></a>}}
\if{latex}{\out{\hypertarget{method-debug}{}}}
\subsection{Method \code{debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_rtext_extended$debug(pos = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ls"></a>}}
\if{latex}{\out{\hypertarget{method-ls}{}}}
\subsection{Method \code{ls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_rtext_extended$ls(what = c("self", "private"), class = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-message"></a>}}
\if{latex}{\out{\hypertarget{method-message}{}}}
\subsection{Method \code{message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_rtext_extended$message(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-warning"></a>}}
\if{latex}{\out{\hypertarget{method-warning}{}}}
\subsection{Method \code{warning()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_rtext_extended$warning(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_rtext_extended$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
