% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext.R, R/rtext_base.R
\docType{class}
\name{rtext}
\alias{rtext}
\alias{rtext_base}
\title{R6 class - linking text and data}
\format{\code{\link{R6Class}} object.}
\usage{
rtext

rtext_base
}
\value{
Object of \code{\link{R6Class}}

Object of \code{\link{R6Class}}
}
\description{
R6 class - linking text and data

R6 class - basic workhorse methods for rtext
}
\keyword{data}

