% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext.R
\docType{class}
\name{rtext}
\alias{rtext}
\title{text class}
\format{\code{\link{R6Class}} object.}
\usage{
rtext
}
\value{
Object of \code{\link{R6Class}}
}
\description{
text class
}
\section{Fields}{

\describe{
\item{\code{text}}{a single character string / character vector of length one}

\item{\code{file}}{path to a file from which text was read in}

\item{\code{encoding}}{encoding to be assumed for the text (will always be UTF-8, because.)}

\item{\code{sourcetype}}{list of logicals that capture where the text came from when
initializing object: text or file.
(text=NULL, file=NULL) : "empty";
(text="", file=NULL | file="") : "text";
(text=NULL, file="") : "file"}
}}
\examples{
mytext <- rtext$new("Hallo World")
mytext$text_get(100)
mytext$char_get()
mytext$text_show()

mytext <- rtext$new(c("Hallo","World"))
mytext$text_get(100)
mytext$text_show()


}
\keyword{data}

