% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_base.R
\docType{class}
\name{rtext_base}
\alias{rtext_base}
\title{rtext_base : basic workhorse for rtext}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}}
}
\description{
rtext_base : basic workhorse for rtext

rtext_base : basic workhorse for rtext
}
\seealso{
\code{\link{rtext}}
}
\keyword{data}
\section{Super class}{
\code{\link[rtext:R6_rtext_extended]{rtext::R6_rtext_extended}} -> \code{rtext_base}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtext_base$new()}}
\item \href{#method-info}{\code{rtext_base$info()}}
\item \href{#method-text_show}{\code{rtext_base$text_show()}}
\item \href{#method-text_get}{\code{rtext_base$text_get()}}
\item \href{#method-text_get_lines}{\code{rtext_base$text_get_lines()}}
\item \href{#method-char_get}{\code{rtext_base$char_get()}}
\item \href{#method-char_add}{\code{rtext_base$char_add()}}
\item \href{#method-char_delete}{\code{rtext_base$char_delete()}}
\item \href{#method-char_replace}{\code{rtext_base$char_replace()}}
\item \href{#method-char_length}{\code{rtext_base$char_length()}}
\item \href{#method-char_data_set}{\code{rtext_base$char_data_set()}}
\item \href{#method-char_data_set_regex}{\code{rtext_base$char_data_set_regex()}}
\item \href{#method-char_data_get}{\code{rtext_base$char_data_get()}}
\item \href{#method-hash_get}{\code{rtext_base$hash_get()}}
\item \href{#method-clone}{\code{rtext_base$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rtext" data-topic="R6_rtext_extended" data-id="debug">}\href{../../rtext/html/R6_rtext_extended.html#method-debug}{\code{rtext::R6_rtext_extended$debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtext" data-topic="R6_rtext_extended" data-id="get">}\href{../../rtext/html/R6_rtext_extended.html#method-get}{\code{rtext::R6_rtext_extended$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtext" data-topic="R6_rtext_extended" data-id="ls">}\href{../../rtext/html/R6_rtext_extended.html#method-ls}{\code{rtext::R6_rtext_extended$ls()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtext" data-topic="R6_rtext_extended" data-id="message">}\href{../../rtext/html/R6_rtext_extended.html#method-message}{\code{rtext::R6_rtext_extended$message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtext" data-topic="R6_rtext_extended" data-id="warning">}\href{../../rtext/html/R6_rtext_extended.html#method-warning}{\code{rtext::R6_rtext_extended$warning()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$new(
  text = NULL,
  text_file = NULL,
  encoding = "UTF-8",
  id = NULL,
  save_file = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text_show"></a>}}
\if{latex}{\out{\hypertarget{method-text_show}{}}}
\subsection{Method \code{text_show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$text_show(
  length = 500,
  from = NULL,
  to = NULL,
  coll = FALSE,
  wrap = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text_get"></a>}}
\if{latex}{\out{\hypertarget{method-text_get}{}}}
\subsection{Method \code{text_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$text_get(length = Inf, from = NULL, to = NULL, split = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text_get_lines"></a>}}
\if{latex}{\out{\hypertarget{method-text_get_lines}{}}}
\subsection{Method \code{text_get_lines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$text_get_lines(length = Inf, from = NULL, to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_get"></a>}}
\if{latex}{\out{\hypertarget{method-char_get}{}}}
\subsection{Method \code{char_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_get(length = Inf, from = NULL, to = NULL, raw = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_add"></a>}}
\if{latex}{\out{\hypertarget{method-char_add}{}}}
\subsection{Method \code{char_add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_add(what = NULL, after = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_delete"></a>}}
\if{latex}{\out{\hypertarget{method-char_delete}{}}}
\subsection{Method \code{char_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_delete(n = NULL, from = NULL, to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_replace"></a>}}
\if{latex}{\out{\hypertarget{method-char_replace}{}}}
\subsection{Method \code{char_replace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_replace(from = NULL, to = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_length"></a>}}
\if{latex}{\out{\hypertarget{method-char_length}{}}}
\subsection{Method \code{char_length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_data_set"></a>}}
\if{latex}{\out{\hypertarget{method-char_data_set}{}}}
\subsection{Method \code{char_data_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_data_set(x = NULL, i = NULL, val = NA, hl = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_data_set_regex"></a>}}
\if{latex}{\out{\hypertarget{method-char_data_set_regex}{}}}
\subsection{Method \code{char_data_set_regex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_data_set_regex(x = NULL, pattern = NULL, val = NA, hl = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_data_get"></a>}}
\if{latex}{\out{\hypertarget{method-char_data_get}{}}}
\subsection{Method \code{char_data_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$char_data_get(from = 1, to = Inf, x = NULL, full = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hash_get"></a>}}
\if{latex}{\out{\hypertarget{method-hash_get}{}}}
\subsection{Method \code{hash_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$hash_get(name = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtext_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
