% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_export.R
\docType{class}
\name{rtext_export}
\alias{rtext_export}
\title{R6 class - linking text and data}
\format{\code{\link{R6Class}} object.}
\usage{
rtext_export
}
\value{
Object of \code{\link{R6Class}}
}
\description{
R6 class - linking text and data
}
\keyword{data}

