% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtext_loadsave.R
\docType{class}
\name{rtext_loadsave}
\alias{rtext_loadsave}
\title{R6 class - load and save methods for rtext}
\format{\code{\link{R6Class}} object.}
\usage{
rtext_loadsave
}
\value{
Object of \code{\link{R6Class}}
}
\description{
R6 class - load and save methods for rtext
}
\keyword{data}

